Module Uno

type color =

| Red

| Green

| Yellow

| Blue

The variant to represent the different colors of the cards

type special =

| Reverse

| Skip

| PlusTwo

The variant to represent the different kinds of special cards

type card =

| Regular of color * int

| Special of color * special

| Wild

| PlacedWild of color

The variant for the card type

type directions =

| Clockwise

| Counterclockwise

The variant for the direction the game will proceed in

The record type for player

type player = {

name : string;

mutable numcards : int;

mutable cards : card list;

curr_card_player : int option;

win : bool option;

}

module type Game = sig ... end

module GameInterface : Game

val create_game : string -> unit